Created an Api call file for the cloudinary url
- Put all the fetch code for the cloudinary url in a api call folder to modularize the code better.
- Removed unesscary component Submit File
- implemented better code considerations
