+++ client/src/components/UploadImageForm/useStyles.tsx
@@ -1,0 +1,14 @@
+import { makeStyles } from '@material-ui/core/styles';
+const useStyles = makeStyles((theme) => ({
+  dropzone: {
+    border: `2px dashed ${theme.palette.primary.main}`,
+    borderRadius: theme.shape.borderRadius,
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    background: theme.palette.background.default,
+    height: theme.spacing(10),
+    outline: 'none',
+  },
+}));
+export default useStyles;
