+++ client/src/interface/Upload.ts
@@ -4,3 +4,7 @@ export interface UploadableFile {
   file: File;
   errors: FileError[];
 }
+
+export interface UploadApiData {
+  cloudinaryUrl: string[];
+}
